Update AllGraph.js
Changes added to fix a bug: When detecting if a change of day, month or year is produced to try to draw a mark, the system accumulates an error of one hour if there is a hour change between min and max dates.

This makes that graphs with dates on an axis that tries to represent data that includes March or October, but not both of them (the error will autocorrect in this cases), doesn't draw the last day/month/year mark and label.

To correct it, the comparisons between current date and max date in days, months and years is parsed to integer values, this rounds any hour deviation and fix the problem. 

More testing is needed to check that this fix doesn't has collateral effects.