Update AllGraph.js
Chrome seems to be unable to locate the correct mouse position once the user has scrolled the page. This seems to be related to different behaviour between browsers in document.body.scrollTop, as the working behaviour is returning always a scroll value of 0 I've removed it entirely from the calculations of the current mouse position.

The bug in Chrome seems to be fixed, no collateral effect seems to happen.
Further test is needed.